+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.15]
+
+### Added
+
+- `remote first` logic for observability platform
+
+### Updated
+
+- minor updates on UI
+- you can read some default values from env vars
+
 ## [0.7.14]
 
 ### Updated